requirements.txt
langchain==0.3.26
langchain-core==0.3.26
flask==3.1.1 
sentence-transformers==4.1.0
pypdf==5.6.1 
python-dotenv==1.1.0
langchain-google-genai==2.1.8
google-generativeai>=0.8.0
langchain-community==0.3.26
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.9
pinecone>=6.0.0,<7.0.0
langchain-pinecone==0.2.8
gunicorn==21.2.0
